# Reduce allocations for generated constructors