For generated types, we often have a public constructor and then internal/base constructors. We reduce the space used when calling one of these public constructors by refactoring the relevant ObjectReference code. Instead of creating and immediately invoking a lambda to give us an object reference, we inline the code in the public constructor. 
 
Also included is a refactoring of the generated "_lazyInterfaces" dictionary. We saw before that it was being initialized in those internal constructors, but without any references to local variables. We move them to the global level and use the Interlocked.CompareExchange pattern of lazy initialization, rather than the .NET Lazy<T> type. 

I have run some benchmarks and have two scenarios. 

For the change to constructors and making/invoking a lambda, we only make this change for net5.0 code. The code to assign _inner is different for netstandard compatibility, making our optimization untenable. However, the change to the lazy interface initializations provides improvements to all runtimes (net5, net3, netstd2).  

**Net5.0**

| Codepath | Master - time | Topic - time | Master - memory | Topic - memory |
| ----------- | --------------  | ------------ | -------------------- | ----------------- |
| Constructors | 3.89 micro | 3.74 micro | 584 B | 392 B |
| Lazy Interface | 11.57 micro | 10.53 micro | 2752 B | 1864 B |

**NetStandard2.0**


| Codepath | Master - time | Topic - time | Master - memory | Topic - memory |
| ----------- | --------------  | ------------ | -------------------- | ----------------- |
| Constructors | 17.59 micro | 17.86 micro | 3481 B | 3453 B |
| Lazy Interface | 10.98 micro | 10.07 micro | 3000 B | 2024 B |