@j0shuams 2022-01-12
Example of difference in generated code, taken from `src\Projections\Windows\GeneratedFiles\net5.0\Windows.Storage.Streams.cs`

**Code generated on master branch**

```csharp
public sealed class DataReader : IDataReader, ...
{
        private IntPtr ThisPtr => _inner == null ? (((IWinRTObject)this).NativeObject).ThisPtr : _inner.ThisPtr;

        private IObjectReference _inner = null;

        private readonly Dictionary<Type, object> _lazyInterfaces;
        
        ...
        
        internal sealed class _IDataReaderFactory { ... }
        
        public DataReader(IInputStream inputStream) : this(((Func<IObjectReference>)(() => {
            IntPtr ptr = (_IDataReaderFactory.Instance.CreateDataReader(inputStream));
            try
            {
                return (ComWrappersSupport.GetObjectReferenceForInterface(ptr));
            }
            finally
            {
                MarshalInspectable<object>.DisposeAbi(ptr);
            }
        }))())
        {
            ComWrappersSupport.RegisterObjectForInterface(this, ThisPtr);
            ComWrappersHelper.Init(_inner, false);
        }
        
        ...
        
        internal DataReader(IObjectReference objRef)
        {
            _inner = objRef.As(GuidGenerator.GetIID(typeof(IDataReader).GetHelperType()));


            _lazyInterfaces = new Dictionary<Type, object>(1)
            {
                {typeof(global::System.IDisposable), new Lazy<global::System.IDisposable>(() => (global::System.IDisposable)(object)new SingleInterfaceOptimizedObject(typeof(global::System.IDisposable), _inner ?? ((IWinRTObject)this).NativeObject))},
            };

        }
        
        ...
        
        private global::System.IDisposable AsInternal(InterfaceTag<global::System.IDisposable> _) =>  ((Lazy<global::System.IDisposable>)_lazyInterfaces[typeof(global::System.IDisposable)]).Value;


}
```

**Code generated on topic branch jlarkin/objref-func**

```csharp
public sealed class DataReader : IDataReader, ...
{
        private IntPtr ThisPtr => _inner == null ? (((IWinRTObject)this).NativeObject).ThisPtr : _inner.ThisPtr;

        private IObjectReference _inner = null;
        
        private volatile global::System.IDisposable _lazy_global__System_IDisposable;

        private global::System.IDisposable Make__lazy_global__System_IDisposable()
        {
            global::System.Threading.Interlocked.CompareExchange(ref _lazy_global__System_IDisposable, (global::System.IDisposable)(object)new SingleInterfaceOptimizedObject(typeof(global::System.IDisposable), _inner ?? ((IWinRTObject)this).NativeObject), null);
            return _lazy_global__System_IDisposable;
        }

        ...
        
        internal sealed class _IDataReaderFactory { ... }
        
        public DataReader(IInputStream inputStream) 
        { 
        IntPtr ptr = (_IDataReaderFactory.Instance.CreateDataReader(inputStream)); 
        try 
        { 
        _inner = ComWrappersSupport.GetObjectReferenceForInterface(ptr); 
        } 
        finally 
        { 
        MarshalInspectable<object>.DisposeAbi(ptr); 
        }

        ComWrappersSupport.RegisterObjectForInterface(this, ThisPtr);
        ComWrappersHelper.Init(_inner, false);
        }
        
        ...
        
        internal DataReader(IObjectReference objRef)
        {
            _inner = objRef.As(GuidGenerator.GetIID(typeof(IDataReader).GetHelperType()));

        }
        
        ... 
        
        private global::System.IDisposable AsInternal(InterfaceTag<global::System.IDisposable> _) =>  Make__lazy_global__System_IDisposable();

```